+++ server/src-exec/Main.hs
@@ -135,6 +135,7 @@ runApp env (HGEOptionsG rci metadataDbUrl hgeCmd) = do
       let defaultSourceConfig = maybeDefaultPgConnInfo <&> \(dbUrlConf, _) ->
             let pgSourceConnInfo = PostgresSourceConnInfo dbUrlConf
                                    (Just setPostgresPoolSettings{_ppsRetries = maybeRetries <|> Just 1})
+                                   False
             in PostgresConnConfiguration pgSourceConnInfo Nothing
       res <- runTxWithMinimalPool _gcMetadataDbConnInfo $ downgradeCatalog defaultSourceConfig opts initTime
       either (printErrJExit DowngradeProcessError) (liftIO . print) res